requirements.txt
appdirs==1.4.4
build==1.0.3
buildozer==1.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Cython==0.29.33
distlib==0.3.8
docutils==0.20.1
filelock==3.13.1
idna==3.6
Jinja2==3.1.3
Kivy==2.3.0
Kivy-examples==2.3.0
Kivy-Garden==0.1.5
kivy-garden.mapview==1.0.6
kivymd==1.2.0
mapview==1.0.6
MarkupSafe==2.1.5
packaging==23.2
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plyer==2.1.0
ptyprocess==0.7.0
Pygments==2.17.2
pyjnius==1.6.1
pyproject_hooks==1.0.0
requests==2.31.0
sh==1.14.3
toml==0.10.2
tomli==2.0.1
urllib3==2.2.1
virtualenv==20.25.0
